A couple tweaks(showing coffee all the time and more)
So I made sure the coffee cup was never hidden during that last sequence.  Before I was playing around with hiding it before he asked for it.  But I think it's better if you just always see it there and just have the option to get it before he asks.  Also added more jibs... and did some other tweaks to the whole thing.