Started implementing CoffeeHit, respawn, and 2nd doom guy
Need to stop, not finished with what I was working on. But this is what I got so far today.  I'll work more tonight.  These are the issues right now: Coffee respawns but you can't use it again, When throwing at boss it's too easy to just hit the desk, in e1m3 sequence you're invincible.