Started testing the posters
Ran into a few texture issues.  Didn't have time to work them out.  See if the textures in the office are working(Whiteboard,NewVent).  The words on the posters are lost, but the point is communicated.  Still works in my opinion.  First I tried making them huge posters, but somehow that didn't feel right.